+++ src/Admin/ReceiptAdmin.php
@@ -103,7 +103,7 @@ final class ReceiptAdmin extends AbstractBaseAdmin
                     'label' => 'backend.admin.invoice.year',
                     'required' => true,
                     'choices' => InvoiceYearMonthEnum::getYearEnumArray(),
-                    'preferred_choices' => [ $currentYear ],
+                    'preferred_choices' => [$currentYear],
                 ]
             )
             ->add(
@@ -190,7 +190,7 @@ final class ReceiptAdmin extends AbstractBaseAdmin
                 ]
             )
         ;
-        if ($this->isFormToCreateNewRecord() || (!$this->isFormToCreateNewRecord() && !$this->getSubject()->getStudent()->isPaymentExempt())) {
+        if ($this->isFormToCreateNewRecord() || (!$this->isFormToCreateNewRecord() && !$this->getSubject()?->getStudent()->isPaymentExempt())) {
             $form
                 ->add(
                     'isSepaXmlGenerated',