+++ CHANGELOG.md
@@ -7,6 +7,7 @@ Changelog
 ##### Version 5.8.04 (2025-03-28)
  * Symfony v6.4.20 update
  * add getUnsubscriptionDate Student entity attribute
+ * update README and LICENSE
 
 ##### Version 5.8.03 (2025-03-19)
  * add more ROLE user cases into testing suite